discord-bot/requirements.txt
opencv-python
discord.py
requests